backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
httpx==0.25.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.1.0,<5.0.0  # Versão compatível com passlib (evita bcrypt 5.x)
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1
python-dotenv==1.0.0
pyufbr==0.1.0
fpdf2==2.7.6
weasyprint==62.3
appwrite>=6.0.0
